Add unit tests for CuratorFactory Assert that the DiscoveryFactory methods are used
Summary:
=======
We'd like some test coverage of the CuratorFactory.
This factory isn't an ideal candidate for unit testing, as it is highly coupled with the Curator library.
Integration tests would provide better certainty here

A result of that high coupling is the number of mocks, particularly for types that discovery doesn't own.
If and when the contracts for curator types change, these mocks may require modification.

That said, a few simple unit tests should provide us with some comfort.
Because the factory methods rely on internal builders, we'll assume that calls to our configuration object are used correctly.

Last but not least, CuratorFactory creates a CuratorManager, which calls CuratorFramework::start.
CuratorFramework::start will attempt to connect to a zookeeper instance as defined by the provided connection string.
To keep the curator from throwing on invalid connection strings or connecting to real zookeeper instances, we will set up a local in-memory test server.
We need to include the curator-test artifact (test scope) and use it's TestingServer class.
TestingServer provides a connection string that will allow the test CuratorFramework to connect to it.

Actions:
=======
* Add curator-test artifact to the pom.
* Add unit tests for curator, relying on TestingServer as our zookeeper "instance."