Initial integration test Register a dropwizard app with the discovery bundle.
Summary:
=======
In order to test features that are more tightly coupled with curator, we will need to use integration tests.

Our first test is to assert that a dropwizard application with the discovery bundle added will register the service.

We will use DropwizardAppRule to set up a trivial dropwizard application with the discovery bundle added.
We will use curator's provided TestingServer to take the place of an actual zookeeper instance.
We will use curator to access the TestingServer to verify that instance metadata has been posted.

We need to share the TestingServer between the bundle and the test.
To do this, we will use a custom Configuration class.
TestApplication will initialize the TestingServer and pass it to the configuration.
The configuration is shared with the test, and the TestingServer can be retrieved there.

Actions:
=======
* Create TestDiscoveryFactory, a subclass of DiscoveryFactory that returns a quorumSpec (connection string) based on a TestingServer's connection string.
* Create TestingServerManager, an implementation of Managed that stops the TestingServer at the end of the app lifecycle.
* Create TestConfiguration to configure the application that integrations tests will run against.
  * TestConfiguration will also hold the TestingServer once the application is running.
* Create TestApplication, a trivial dropwizard application with the discovery bundle initialized/bootstrapped.
  * TestApplication initializes the TestingServer and passes it to the TestConfiguration.
  * TestApplication also registers a TestingServerManager for TestingServer, to ensure it is properly shut down after the test.
* Create test-config.yaml, a baseline configuration to initialize TestConfiguration.
* Create DiscoveryBundleIntegrationTest.
* Add a DropwizardAppRule to DiscoveryBundleIntegrationTest to initialize a TestApplication with test-config.yaml.
* Add a test that uses curator to assert that the bundle has registered an instance of a service.
  * Initialize a CuratorFramework with the TestingServer from TestConfiguration.
  * Search for the node under the expected namespace, base-path, and service name.
  * Use the node name to retrieve the serialized ServiceInstance.
  * Deserialize the instance and assert that the retrieved object is in-tact.